Player IDs are uint32s, not uint16s. Also, now drawing metaserver users' colors in swatches by their names in the SDL UI.  The trick is, the swatches aren't yet hooked up to the players' actual colors.  But it's getting closer.  :)
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@2122 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Network/network_dialogs_sdl.cpp
@@ -839,7 +839,7 @@ public:
 	{
 	}
 
-	void receivedChatMessage(const std::string& senderName, uint16 senderID, const std::string& message)
+	void receivedChatMessage(const std::string& senderName, uint32 senderID, const std::string& message)
 	{
 		m_chatHistory.append_chat_entry(senderName.c_str(), 0xaaaaaaaa, 0xaaaaaaaa, message.c_str());
 		m_chatHistory.get_owning_dialog()->draw_dirty_widgets();
